Add files via upload
Bagging predictors is a method for generating multiple versions of a predictor and using these to get
an aggregated predictor. The aggregation averages over the versions when predicting a numerical outcome and
does a plurality vote when predicting a class. The multiple versions are formed by making bootstrap replicates of
the learning set and using these as new learning sets. Tests on real and simulated data sets using classification and
regression trees and subset selection in linear regression show that bagging can give substantial gains in accuracy.
The vital element is the instability of the prediction method. If perturbing the learning set can cause significant
changes in the predictor constructed, then bagging can improve accuracy.