Add files via upload
Special emphasis is given to estimating potentially complex parametric or nonparametric models, including generalized linear and additive models as well as regression
models for survival analysis. Concepts of degrees of freedom and corresponding Akaike or Bayesian information criteria, particularly useful for regularization and variable selection in high-dimensional covariate spaces,are discussed as well.